upgrade to astro 5
