+++ src/components/Meta.astro
@@ -1,5 +1,5 @@
 ---
-import { ViewTransitions } from "astro:transitions";
+import { ClientRouter } from "astro:transitions";
 
 interface Props {
   title: string;
@@ -59,4 +59,4 @@ const { title, description, content = "" } = Astro.props;
 <meta name="generator" content={Astro.generator} />
 
 <title>{title}</title>
-<ViewTransitions />
+<ClientRouter />
