+++ src/content/slides/slides.json
@@ -1,9 +1,11 @@
 [
     {
+        "id": 2,
         "description": "Nov. 2022 - delivered a talk on Data Analysis in Io(B)T at a DRDO Sponsored One Week Workshop On 'Blockchain Assisted Secured Data Analytics In Internet Of Battlefield Things (ioBt)'",
         "link": "files/iobt.pdf"
     },
     {
+        "id": 1,
         "description": "Nov. 2023 - delivered a talk on AI in IoT at a ATAL Sponsored Faculty Development Programme on (Artificial Intelligence and Machine Learning for Industrial Internet of Things 5.0)",
         "link": "files/llm.pdf"
     }
